// manager_frontend/src/client/models/ManagerRepairComplaintPresetCreatePayload.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ManagerRepairComplaintPresetCreatePayload = {
    complaint_group?: string;
    customer_phrase: string;
    document_wording?: string;
    likely_diagnosis?: string;
    is_favorite?: boolean;
    is_active?: boolean;
    sort_order?: number;
    comment?: (string | null);
};