// manager_frontend/src/client/models/ManagerRepairComplaintPresetResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ManagerRepairComplaintPresetResponse = {
    id: number;
    complaint_group?: string;
    customer_phrase: string;
    document_wording?: string;
    likely_diagnosis?: string;
    is_favorite?: boolean;
    is_active?: boolean;
    sort_order?: number;
    comment?: (string | null);
    created_at: string;
    updated_at?: (string | null);
};