// manager_frontend/src/client/models/ManagerRepairComplaintPresetUpdatePayload.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ManagerRepairComplaintPresetUpdatePayload = {
    complaint_group?: (string | null);
    customer_phrase?: (string | null);
    document_wording?: (string | null);
    likely_diagnosis?: (string | null);
    is_favorite?: (boolean | null);
    is_active?: (boolean | null);
    sort_order?: (number | null);
    comment?: (string | null);
};